requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
albumentations>=1.3.0
tqdm>=4.62.0
PyYAML>=6.0.0
Pillow>=8.3.0
